requirements.txt
Flask==3.0.3
flask-cors==4.0.1
SQLAlchemy==2.0.27
python-dotenv==1.0.1
openai==0.28.1
pymysql==1.1.0
pytest==8.0.2
python-docx==1.1.2
PyPDF2==3.0.1
pdfplumber==0.11.4
regex==2024.9.11